postgres-data-modeling/requirements.txt
psycopg2-binary
pandas
numpy
jupyter
ipython-sql